% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Post_Prob.R
\name{Bayesian_Post_Prob}
\alias{Bayesian_Post_Prob}
\title{Bayesian posterior probability P(Theta>cutoff|data;prior)}
\usage{
Bayesian_Post_Prob(n, succ, beta_par, cutoff)
}
\arguments{
\item{n}{number of patients for which results are available}

\item{succ}{number of successes}

\item{beta_par}{two shape parameters c(alpha,beta) for beta distribution}

\item{cutoff}{Proportion corresponding with some hypothesis, e.g. null hypothesis}
}
\description{
This function calculates the Bayesian posterior probability P(Theta>cutoff|data;prior)
}
\examples{
\dontrun{
Bayesian_post_prob(n=68,succ=52,beta_par=c(1,1),cutoff=0.66)
}

}
