% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_binomial.R
\name{Beta_binomial}
\alias{Beta_binomial}
\title{Beta-binomial distribution of second-stage binary outcomes}
\usage{
Beta_binomial(N, n1, x1, beta_par)
}
\arguments{
\item{N}{total number of patients (stage1 + stage2)}

\item{n1}{total number of patients in stage1}

\item{x1}{number of successes in stage1}

\item{beta_par}{two shape parameters c(alpha,beta) for beta distribution}
}
\value{
a dataframe with two variables: 
\itemize{
\item x2: all possible outcomes (number of successes) for stage2
\item prediction: density for each outcome
}
}
\description{
This function calculates the probability density for second-stage binary outcomes
}
\examples{
\dontrun{
Beta_binomial(N=40,n1=23,x1=16,beta_par=c(1,1))
}
}
