% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContMon_getbound.R
\name{ContMon_getbound}
\alias{ContMon_getbound}
\title{Get boundaries for continuous trial monitoring}
\usage{
ContMon_getbound(
  SS,
  target,
  Beta_dis,
  P_target = NULL,
  maxtox = NULL,
  prt1 = 0,
  tox_ass = seq(0.1, 0.5, 0.05),
  sim = 0,
  nsim,
  out
)
}
\arguments{
\item{SS}{total sample size, e.g. calculated for primary efficacy endpoint}

\item{target}{target toxicity rate}

\item{Beta_dis}{two parameters Beta(alpha,beta) of the prior Beta distribution}

\item{P_target}{P(toxicity rate>target|data)>=P_target as decision threshold}

\item{maxtox}{maximum allowable observed toxicity rate, where boundary not crossed}

\item{prt1}{value for number of patients in starting cohort where alternative safety rule is followed, e.g. 3+3}

\item{tox_ass}{vector of rates for calculating Bayesian probabilities P(toxicity rate<=tox_ass[]|boundary crossed) and frequentist
operating characteristics}

\item{sim}{binary indicator: simulations for frequentist operating characteristics (0=NO; 1=YES)}

\item{nsim}{number of simulated datasets}

\item{out}{path for storing csv output files}
}
\description{
This function calculates boundaries for every sample size
}
\details{
Two parameters should be specified: either 'target' and 'maxtox'
    either 'target' and 'P_target'. If the Bayesian posterior probability, given the data
    is higher than 'P_target', i.e. P(toxicity rate>target|c+1,n)>=P_target, the boundary is crossed.
    For every 'n', the boundary 'c' is given, which corresponds with the maximum allowed
    toxicity. So the boundary is crossed at 'c+1'

P_Bayes= Bayesian posterior probability P(toxicity rate>target|c,n)

P_obs= c/n or 'allowable proportion': if number of events <=c then continuation, if c+1: boundary crossed

'Bayes_result' gives the Bayesian posterior probability P(toxicity rate<=target|c+1,n), so given that
    this boundary 'c+1' is exceeded, for assumptions, specified by the 'tox_ass' parameter

Prob(X>n|n,tox_true=...) gives the Prob(X>=c+1|different assumptions ('toxass'))

Cum Prob(...) gives the cumulative probability of at least 1x/2x/3x and 4x crossing the boundary
    by n patients
}
\examples{
\dontrun{
ContMon_getbound(SS=19,target=0.2,Beta_dis=c(1,1),P_target=NULL,maxtox=0.33,
    tox_ass=seq(0.10,0.5,0.05),sim=1,nsim=100,prt1=6,out=NULL)
}

}
