% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContMon_plotbound.R
\name{ContMon_plotbound}
\alias{ContMon_plotbound}
\title{Plot properties of continuous safety monitoring boundaries}
\usage{
ContMon_plotbound(data_in, out)
}
\arguments{
\item{data_in}{resulting list (containing data with results, and vector with assumptions for calculating
properties) from the 'getbound' function}

\item{out}{path for storing png output files}
}
\description{
This function plots the properties of continuous safety monitoring boundaries
}
\details{
Freq_prob_date gives the P(X>c+1|n,tox_true) for different n and for different assumptions,
    specified by the 'tox_ass' parameter

Bayes_probgives the Bayesian posterior probability P(toxicity rate<=target|c+1,n), so given that
    this boundary 'c+1' is exceeded, for different assumptions, specified by the 'tox_ass' parameter

Cum_Freq_prob_1 gives the cumulative probability of at least 1x crossing the boundary

Cum_Freq_prob_2 gives the cumulative probability of at least 1x/2x/3x and 4x crossing the boundary
}
\examples{
\dontrun{
result<-getbound(SS=39,target=0.2,P_target=0.9,maxtox=NULL,tox_ass=seq(0.10,0.5,0.05),
    sim=1,nsim=100,prt1=6,out=NULL)
ContMon_plotbound(data_in=result,out="C:/Users/kdhollander/Downloads/")
}

}
