% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhII_Freq_Prop.R
\name{getPPdesign}
\alias{getPPdesign}
\title{PhII Bayesian design with predictive probability}
\usage{
getPPdesign(
  N,
  Delta_T,
  interim_type,
  interim,
  cohortsize = NULL,
  interimstart = 10,
  interimstop = 5,
  Delta_L = 0,
  Delta_U = 1,
  p0,
  pa,
  Beta_dis,
  Freqprop,
  nsim
)
}
\arguments{
\item{N}{this must be a vector: of different total sample sizes. Different N's can be investigated}

\item{Delta_T}{This must be vector: of different threshold values, defining the posterior \cr
probability P(Delta>=p0|data at the final analysis), for a succesful trial.}

\item{interim_type}{choice between 3 character types: "Perc", "cont" or "fix". \cr
Only one sequence at once can be investigated: \cr
1) "Perc": percentage(s) of information, e.g. c(0.5,0.75), specified as a vector via the 'interim' parameter \cr
2) "fix": fixed number of patients, e.g. c(10,20), specified as a vector via the 'interim' parameter \cr
3) "cont": continuous monitoring. Run-in without monitoring (e.g. first 10) is specified in the "interimstart" \cr
parameter. Run-out without monitoring (e.g. last 5) is specified in the "interimstop" parameter.}

\item{interim}{This must be a vector, only to be specified when the parameter 'interim type' is "Perc" or "fix" \cr
e.g. c(0.5,0.75) when 'interim type' is "Perc", or e.g. c(10,20) when 'interim type' is "fix"}

\item{cohortsize}{This must be a scalar, indicating the frequency of continuous monitoring. \cr
Only needs to be defined in case the parameter 'interim_type'="cont"}

\item{interimstart}{This must be a scalar, specifying run-in without monitoring. \cr
Only needs to be defined in case the parameter 'interim_type'="cont". Default is 10}

\item{interimstop}{This must be a scalar, specifying run-out without monitoring. \cr
Only needs to be defined in case the parameter 'interim_type'="cont". Default is 5.}

\item{Delta_L}{Vector for futility ("Lower") boundary. This may be a scalar, if the parameter "interim"='cont' \cr
or if there is only one interim analysis. This corresponds with a threshold value for the predictive probability \cr
of a succesful result at the final analysis (Posterior probability >=Delta_T). Default value=0 (no interim for futility).}

\item{Delta_U}{Vector for efficacy ("Upper") boundary. This may be a scalar, if the parameter "interim"='cont' \cr
or if there is only one interim analysis. This corresponds with a threshold value for the predictive probability \cr
of a succesful result at the final analysis (Posterior probability >=Delta_T). Default value=1 (no interim for efficacy).}

\item{p0}{probability of the uninteresting response (null hypothesis \eqn{H0})}

\item{pa}{probability of the interesting response (alternative hypothesis Ha)}

\item{Beta_dis}{two parameters Beta(alpha,beta) of the prior Beta distribution}

\item{Freqprop}{"pow" for power (simulation under Ha) or "typeI" for simulation under \eqn{H0}}

\item{nsim}{number of simulations}
}
\value{
a data.frame with elements
\itemize{
\item N: Different sample sizes for which frequentist properties simulated
\item Delta_T: Different thresholds for which frequentist properties simulated
\item Boundaries: Different boundaries (combination of Delta_L and Delta_U) for which frequentist properties are simulated
\item N_interim: number of interim analyses (final analysis not included)
\item pow: proportion of simulations where \eqn{H0} was rejected
\item eff_fin: proportion of simulations where \eqn{H0} was rejected only at the final analysis
\item eff_stop: proportion of simulations where \eqn{H0} was rejected at an interim analysis
\item fut_stop: proportion of simulations where trial stopped for futility at an interim analysis
\item int_nrx: number of patients at x'th interim analysis
\item effx: proportion of simulations where trial stopped for efficacy at xth interim analysis
\item futx: proportion of simulations where trial stopped for futility at xth interim analysis
}
}
\description{
A flexible PhII, signle-arm study design, based on Bayesian predictive probability \cr
The frequentist properties can be investigated of combinations of \cr
1) Different sample sizes
2) Different Delta_T
3) Different boundary (efficacy boundary Delta_U and futility boundary Delta_L are linked \cr
and must be specified as a pair of boundaries. So the parameters Delta_U and Delta_L \cr
should have exactly the same structure.
}
\examples{
\donttest{
## Example 1
result<-getPPdesign(N=seq(80,100,5),Delta_T=c(0.95),Delta_L=list(c(0.2,0.2)),
Delta_U=list(c(0.8,0.8)),interim_type="Perc",interim=c(0.5,0.7),
p0=0.4,pa=0.6,Beta_dis=c(0.4,0.6),Freqprop="pow",nsim=100)
                    
## Example 2 
result<-getPPdesign(N=seq(80,100,5),Delta_T=c(0.95),Delta_L=0.3,Delta_U=1,
interim_type="cont",cohortsize=3,
p0=0.4,pa=0.6,Beta_dis=c(0.4,0.6),Freqprop="pow",nsim=100)

## Example 3 
result<-getPPdesign(N=c(68),Delta_T=c(0.95),interim_type="fix",
interim=c(30,40),Delta_L=list(c(0.1,0.05),c(0.1,0.1)),Delta_U=list(c(1,1),c(1,1)),
p0=0.66,pa=0.8,Beta_dis=c(0.66,0.33),Freqprop="pow",nsim=10)

## Example 4 
result<-getPPdesign(N=c(68),Delta_T=c(0.95),interim_type="fix",
interim=c(30,40),Delta_L=list(c(0.1,0.05),c(0.1,0.1)),Delta_U=list(c(1,1),c(1,1)),
p0=0.66,pa=0.8,Beta_dis=c(0.66,0.33),Freqprop="typeI",nsim=10)

## Example 5 
result<-getPPdesign(N=c(68),Delta_T=c(0.95),interim_type="fix",
interim=c(30),Delta_L=list(c(0.1)),Delta_U=list(c(1,1)),
p0=0.66,pa=0.8,Beta_dis=c(0.66,0.33),Freqprop="pow",nsim=10)

result<-getPPdesign(N=c(68),Delta_T=c(0.95),interim_type="fix",
interim=c(30),Delta_L=list(c(0.1)),Delta_U=list(c(1,1)),
p0=0.66,pa=0.8,Beta_dis=c(0.66,0.33),Freqprop="typeI",nsim=10) 
}
}
\references{
Lee JJ, Liu DD.A predictive probability design for phase II cancer clinical trialsClinical Trials 2008; 5: 93–106
}
