% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_Prob.R
\name{Post_Prob}
\alias{Post_Prob}
\title{Bayesian posterior probability P(Theta>cutoff|data;prior)}
\usage{
Post_Prob(n, x, beta_par, p)
}
\arguments{
\item{n}{number of patients for which results are available}

\item{x}{number of successes}

\item{beta_par}{two shape parameters c(alpha,beta) for beta distribution}

\item{p}{Proportion corresponding with some hypothesis, e.g. null hypothesis}
}
\description{
This function calculates the Bayesian posterior probability P(Theta>cutoff|data;prior)
}
\examples{
\dontrun{
Post_Prob(n=68,x=52,beta_par=c(1,1),p=0.66)
}

}
