% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_Prob_R.R
\name{Post_Prob_R}
\alias{Post_Prob_R}
\title{Bayesian posterior probability, given an observed difference of 2 proportions}
\usage{
Post_Prob_R(
  p_exp,
  p_ctrl,
  n_exp,
  n_ctrl,
  distrisize,
  D_cut,
  beta_par_exp,
  beta_par_ctrl
)
}
\arguments{
\item{p_exp}{observed proportion experimental arm}

\item{p_ctrl}{observed proportion control arm}

\item{n_exp}{number of patients in experimental arm (scalar)}

\item{n_ctrl}{number of patients in control arm (scalar)}

\item{distrisize}{Size of sampled distributions (the larger, the better)}

\item{D_cut}{True difference between two proportions (can be a vector)}

\item{beta_par_exp}{two shape parameters c(alpha,beta) for prior beta distribution experimental arm (scalar)}

\item{beta_par_ctrl}{two shape parameters c(alpha,beta) for prior beta distribution control arm (scalar)}
}
\description{
This function calculates the Bayesian posterior probability 
P(Delta>Dcut|n,x) with Delta=Delta_exp-Delta_ctrl, with Delta_exp~beta(beta_par_exp[1]+x,beta_par_exp[2]+n_exp-x_exp) 
with prior beta distribution beta_par, and Delta_ctrl~beta(beta_par_ctrl[1]+x,beta_par_ctrl[2]+n_ctrl-x_ctrl)
}
\examples{
Post_Prob_R(p_exp=0.475,p_ctrl=0.475-5/50,n_exp=50,n_ctrl=50,distrisize=10^3,
D_cut=c(0,0.05,0.075,0.1),beta_par_exp=c(1,1),beta_par_ctrl=c(1,1))
}
