% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred_Prob.R
\name{Pred_Prob}
\alias{Pred_Prob}
\title{Posterior predictive probability of a successful trial}
\usage{
Pred_Prob(N, n1, x1, beta_par, p, Delta_T)
}
\arguments{
\item{N}{total number of patients (stage1 + stage2)}

\item{n1}{total number of patients in stage1}

\item{x1}{number of successes in stage1}

\item{beta_par}{two shape parameters c(alpha,beta) for beta distribution}

\item{p}{Proportion corresponding with some hypothesis, e.g. null hypothesis}

\item{Delta_T}{Threshold for succesful trial in terms of Bayesian posterior probability
P(theta>p|x1+x2)>Delta_T}
}
\value{
Probability of a succesful trial
}
\description{
Posterior predictive probability of a successful trial, given number of successes
    at the end of stage1
}
\examples{
\dontrun{
Pred_Prob(N=68,n1=38,x1=27,beta.par=c(1  ,1)  ,p=0.664,Delta_T=0.95)
Pred_Prob(N=40,n1=23,x1=16,beta.par=c(0.6,0.4),p=0.6  ,Delta_T=0.9 )
}

}
\references{
Lee JJ, Liu DD.A predictive probability design for phase II cancer clinical trialsClinical Trials 2008; 5: 93–106
}
