% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_Prob.R
\name{Post_Prob}
\alias{Post_Prob}
\title{Bayesian posterior probability, given 1 observed proportion}
\usage{
Post_Prob(n, x, beta_par, Dcut)
}
\arguments{
\item{n}{number of patients}

\item{x}{number of successes}

\item{beta_par}{two shape parameters c(alpha,beta) for beta distribution}

\item{Dcut}{Proportion corresponding with some hypothesis}
}
\description{
This function calculates the Bayesian posterior probability 
P(Delta>Dcut|n,x) ~beta(beta_par[1]+x,beta_par[2]+n-x) with prior beta distribution beta_par
}
\examples{
\dontrun{
Post_Prob.f(n=40,x=16,beta_par=c(0.6,0.4),Dcut=0.6) # Example Lee 2008, p97, Table 1 
(first line of Panel B: Y=0)
}

}
